+++ src/messaging/Crypto.ts
@@ -13,29 +13,6 @@ import { hexToU8a, stringToU8a, u8aToHex, u8aToString } from '@polkadot/util'
 import type { IEncryptedMessage, IEncryptedMessageContents, IMessage } from '../types'
 
 /**
- * Checks if the message object is well-formed.
- *
- * @param message The message object.
- */
-export function verifyMessageEnvelope(message: IMessage): void {
-  const { messageId, createdAt, receiver, sender, receivedAt, inReplyTo } = message
-  if (messageId !== undefined && typeof messageId !== 'string') {
-    throw new TypeError('Message id is expected to be a string')
-  }
-  if (createdAt !== undefined && typeof createdAt !== 'number') {
-    throw new TypeError('Created at is expected to be a number')
-  }
-  if (receivedAt !== undefined && typeof receivedAt !== 'number') {
-    throw new TypeError('Received at is expected to be a number')
-  }
-  Did.validateUri(sender, 'Did')
-  Did.validateUri(receiver, 'Did')
-  if (inReplyTo && typeof inReplyTo !== 'string') {
-    throw new TypeError('In reply to is expected to be a string')
-  }
-}
-
-/**
  * Symmetrically decrypts the result of [[encrypt]].
  *
  * @param encrypted The encrypted message.
