+++ src/messaging/Message.test.ts
@@ -35,8 +35,13 @@ import {
   makeSigningKeyTool,
 } from '../tests'
 import { fromBody, verifyRequiredCTypeProperties } from './utils'
-import { decrypt, encrypt, verifyMessageEnvelope } from './MessageEnvelope'
-import { ensureOwnerIsSender, verify, verifyMessageBody } from './CredentialApiMessageType'
+import { decrypt, encrypt } from './Crypto'
+import {
+  ensureOwnerIsSender,
+  assertKnownMessage,
+  verifyMessageBody,
+  verifyMessageEnvelope,
+} from './CredentialApiMessageType'
 import type {
   IEncryptedMessage,
   IMessage,
@@ -133,7 +138,7 @@ describe('Messaging', () => {
     const decryptedMessage = await decrypt(encryptedMessage, bobEncKey.decrypt, { resolveKey })
     expect(JSON.stringify(message.body)).toEqual(JSON.stringify(decryptedMessage.body))
 
-    expect(() => verify(decryptedMessage)).not.toThrow()
+    expect(() => assertKnownMessage(decryptedMessage)).not.toThrow()
 
     const encryptedMessageWrongContent = JSON.parse(
       JSON.stringify(encryptedMessage)
