+++ src/quote/Error.ts
@@ -7,10 +7,6 @@
 
 export declare class QuoteError extends Error {}
 
-export declare class HashMalformedError extends QuoteError {
-  constructor(hash?: string, type?: string)
-}
-
 export declare class QuoteUnverifiableError extends QuoteError {}
 export declare class SignatureUnverifiableError extends QuoteError {}
 export declare class DidSubjectMismatchError extends QuoteError {}
