+++ src/quote/index.ts
@@ -5,5 +5,6 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-export * from './Quote.js'
-export { QuoteSchema } from './QuoteSchema.js'
+export * from './Quote'
+export { QuoteSchema } from './QuoteSchema'
+export * from './Error'
