+++ src/utils/TypeGuards.ts
@@ -1,4 +1,12 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+
+/**
+ * Copyright (c) 2018-2023, BOTLabs GmbH.
+ *
+ * This source code is licensed under the BSD 4-Clause "Original" license
+ * found in the LICENSE file in the root directory of this source tree.
+ */
+
 import { IAttestation, ICredentialPresentation, IRequestCredentialContent } from '@kiltprotocol/types'
 
 import type {
