+++ src/wellKnownDidConfiguration/wellKnownDidConfiguration.test.ts
@@ -1,3 +1,10 @@
+/**
+ * Copyright (c) 2018-2023, BOTLabs GmbH.
+ *
+ * This source code is licensed under the BSD 4-Clause "Original" license
+ * found in the LICENSE file in the root directory of this source tree.
+ */
+
 import {
   KiltKeyringPair,
   DidUri,
